Fix saml.active flag
Also remove commented out code - has been commented out since 'Initial
Commit'.

Broke the active check into one for `conf` and one for
`conf.saml.active` so as to provide greater feedback possibilities for
users. As it is, if the `!conf` situation arises, it would mean the
default config is also missing from the plugin - so maybe this is not
required, but kept it as per the original fix.

Based on original fix @ fd0150cd2ced1b7482e7907be46a2f21eb024e20

Issue: #33