Remove incorrect spring security config for domain classes
The default values for the domain classes had been overridden but are
no longer correct. So removing to use defaults which are correct.

Discovered while writing new integration spec - which is how I've also
confirmed.

This is only in use anyway for local tests - no impact for users of the
plugin.