Utilise new local role retrieval method in loadUserBySAML and refactor
Feedback in PR #44 highlighted that a block of code could now be replace
with one of the new methods I provided in the previous commit. So this
code mostly fulfills that.

However in doing so, I revisited my new methods and decided they could
be written better - especially if to be used elsewhere. So I refactored
them at the same time.

I've run `./gradlew clean check` and all passes. Plus I've installed the
plugin locally and all still works.

Issue: #43